Created a new ASP .NET Visual Studio Solution using .NET Core 3.1.
The plan for this project is to have an "Admin Page" allowing for viewing of the database from the web (locally) such as viewing, creating, editing, and deleting the tables.

Otherwise, I will create simple form that would act as a "User" page so that an author may post a blog.

The Database will be created within the same solution in Visual Code.