Final Prototype (6/24/2021)
This is the final prototype that I plan to turn in. I would have liked to add and change more but I didn't have the time to get things exactly as I would have liked due to lack of time and this being the first time using ASP .NET C# MVC

Thank you!